Commands put into Command design pattern
