+++ jambl/src/main/java/jambl/Controller/Commands/Command1.java
@@ -1,0 +1,6 @@
+package jambl.Controller.Commands;
+public interface Command1 {
+
+    void execute();
+    
+}
