+++ jambl/src/main/java/jambl/Controller/Commands/Command2.java
@@ -1,0 +1,10 @@
+
+package jambl.Controller.Commands;
+
+public interface Command2 {
+
+    void execute();
+
+    void unexecute();
+    
+}
