+++ jambl/src/main/java/jambl/Controller/Commands/HelpCommand.java
@@ -1,0 +1,15 @@
+package jambl.Controller.Commands;
+
+import jambl.View.View;
+
+public class HelpCommand implements Command1{
+    View view = new View();
+    public HelpCommand(){
+
+    }
+
+    public void execute(){
+        view.printHelp();
+    }
+    
+}
